created Fetch module for Meals
